requirements.txt
pandas~=2.2.2
numpy~=1.25.1
pillow~=10.4.0
opencv-python~=4.10.0.84